Adds test to this API.
Version update.

Supports Tagging API for parameter group, so customers could update tags for their parameter groups after creation.

Moves updateTags function from post_set to pre_build for better format. Deletes and fixes some small mistakes.

Fixs code based on comments from code review.

Fixs possible panic code for some specific cases, and starts doing automated test.
