+++ apis/v1alpha1/ack-generate-metadata.yaml
@@ -1,13 +1,13 @@
 ack_generate_info:
-  build_date: "2022-09-12T19:52:33Z"
-  build_hash: 2944c8772f216656d84ee02d392eaca501274c1e
-  go_version: go1.17.5
-  version: v0.20.1
-api_directory_checksum: 93069d6d43fb053a96eda1dfc900fbe410ce5b1e
+  build_date: "2022-11-29T01:22:28Z"
+  build_hash: c1f00b512eb25d054b6b5faa67674d9a08e28b38
+  go_version: go1.19
+  version: v0.20.1-2-gc1f00b5-dirty
+api_directory_checksum: a1e396caca4bdd1612fa7d09f0ee56f3e4976ff7
 api_version: v1alpha1
 aws_sdk_go_version: v1.44.93
 generator_config_info:
-  file_checksum: 24344616e89a51daae2b05f38438010f32995a4f
+  file_checksum: c20e176b6706e0a4c8db79814d1a1672ad760c81
   original_file_name: generator.yaml
 last_modification:
   reason: API generation
