+++ apis/v1alpha1/generator.yaml
@@ -258,9 +258,9 @@ resources:
       sdk_read_many_post_set_output:
         template_path: hooks/parametergroup/sdk_read_many_post_set_output.go.tpl
       sdk_create_post_set_output:
-        template_path: hooks/parametergroup/sdk_read_many_post_set_output.go.tpl
+        template_path: hooks/parametergroup/sdk_create_post_set_output.go.tpl
       sdk_update_post_set_output:
-        template_path: hooks/parametergroup/sdk_read_many_post_set_output.go.tpl
+        template_path: hooks/parametergroup/sdk_update_post_set_output.go.tpl
       sdk_update_pre_build_request:
         template_path: hooks/parametergroup/sdk_update_pre_build_request.go.tpl
     renames:
