+++ apis/v1alpha1/subnet_group.go
@@ -24,9 +24,9 @@ import (
 //
 // Represents the output of one of the following operations:
 //
-//    * CreateSubnetGroup
+//   - CreateSubnetGroup
 //
-//    * UpdateSubnetGroup
+//   - UpdateSubnetGroup
 //
 // A subnet group is a collection of subnets (typically private) that you can
 // designate for your clusters running in an Amazon Virtual Private Cloud (VPC)
