+++ apis/v1alpha1/types.go
@@ -298,9 +298,9 @@ type Subnet struct {
 
 // Represents the output of one of the following operations:
 //
-//    * CreateSubnetGroup
+//   - CreateSubnetGroup
 //
-//    * UpdateSubnetGroup
+//   - UpdateSubnetGroup
 //
 // A subnet group is a collection of subnets (typically private) that you can
 // designate for your clusters running in an Amazon Virtual Private Cloud (VPC)
