+++ pkg/resource/parameter_group/sdk.go
@@ -118,6 +118,13 @@ func (rm *resourceManager) sdkFind(
 	}
 
 	rm.setStatusDefaults(ko)
+	resourceARN := (*string)(ko.Status.ACKResourceMetadata.ARN)
+	tags, err := rm.getTags(ctx, *resourceARN)
+	if err != nil {
+		return nil, err
+	}
+	ko.Spec.Tags = tags
+
 	ko, err = rm.setParameters(ctx, ko)
 
 	if err != nil {
@@ -260,11 +267,18 @@ func (rm *resourceManager) sdkUpdate(
 	}()
 	if delta.DifferentAt("Spec.ParameterNameValues") {
 		ko, err := rm.resetParameterGroup(ctx, desired, latest)
-
 		if ko != nil || err != nil {
 			return ko, err
 		}
 	}
+
+	if delta.DifferentAt("Spec.Tags") {
+		err = rm.updateTags(ctx, desired, latest)
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	input, err := rm.newUpdateRequestPayload(ctx, desired)
 	if err != nil {
 		return nil, err
