+++ templates/hooks/parametergroup/sdk_create_post_set_output.go.tpl
@@ -1,0 +1,5 @@
+ko, err = rm.setParameters(ctx, ko)
+
+if err != nil {
+    return nil, err
+}
