+++ templates/hooks/parametergroup/sdk_update_pre_build_request.go.tpl
@@ -1,7 +1,13 @@
 if delta.DifferentAt("Spec.ParameterNameValues") {
     ko, err := rm.resetParameterGroup(ctx, desired, latest)
-
     if ko != nil || err != nil {
         return ko, err
+    }
+}
+
+if delta.DifferentAt("Spec.Tags") {
+    err = rm.updateTags(ctx, desired, latest)
+    if err != nil {
+        return nil, err
     }
 }
